Fix plugin and groovy initializer loading
In order to ensure that we can load configuration from downstream kustomizations correctly, we need to jump a few hoops. Jenkins will not discover groovy files nested within the init.groovy.d folder, so we iterate the files we have found and create symlinks for them into init.groovy.d.

Similarly for plugins, we need to concatenate the plugin lists from downstream and feed them into the jenkins-plugin-cli tool.