Adds kubernetes-credentials-provider to base config
This change sets up Kubernetes Secret reading in the base configuration. This is done as it allows for a much nicer Secrets integration between Jenkins and Kubernetes in downstream configurations, without assuming an actual Secrets provider beyond "Kubernetes". How Secrets get into a Kubernetes namespace is left up to the user.